Separate pages into a pages folder.
