+++ src/pages/Calculator/index.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Navigation } from '../../core'
+
+function App() {
+  return (
+    <div className="App">
+      <Navigation />
+      <div><h1>Welcome to the Home Page</h1></div>
+    </div>
+  )
+}
+
+export default App
